Moved render code to game_render.c
