+++ DragonQuestinoWinDev/DragonQuestinoWinDev/DragonQuestinoWinDev.vcxproj
@@ -41,6 +41,7 @@
     <ClCompile Include="..\..\DragonQuestino\game.c" />
     <ClCompile Include="..\..\DragonQuestino\game_data.c" />
     <ClCompile Include="..\..\DragonQuestino\game_items.c" />
+    <ClCompile Include="..\..\DragonQuestino\game_render.c" />
     <ClCompile Include="..\..\DragonQuestino\input.c" />
     <ClCompile Include="..\..\DragonQuestino\math.c" />
     <ClCompile Include="..\..\DragonQuestino\menu.c" />
